commit the graph of showing set range

+++ app/src/main/java/com/example/jetpackcomposeevoluznsewingmachine/ModalClass/DailySummary.kt
@@ -4,6 +4,6 @@ data class DailySummary(
     val day: String,
     val avg_temperature: Double,
     val avg_vibration:Double,
-    val avg_oilLevel:Int
+    val avg_oilLevel:Double
 )
